System.Date.Cache

Description

Formatting time is slow. This package provides mechanisms to cache formatted date.

Types

data DateCacheConf t

Constructors

DateCacheConf

Fields

getTime :: IO t

A function to get a time. E.g epochTime and getCurrentTime.

formatDate :: t -> IO ByteString

A function to format a time.

type DateCacheGetter = IO ByteString

type DateCacheCloser = IO ()

Date cacher

ondemandDateCacher :: Eq t => DateCacheConf t -> IO (DateCacheGetter, DateCacheCloser)

Date cacher which gets a time and formatted it only when returned getter is executed.

clockDateCacher :: Eq t => DateCacheConf t -> IO (DateCacheGetter, DateCacheCloser)

Date cacher which gets a time and formatted it every second. This returns a getter.